Clip boundaries: exact half-open activity in every mode (supersedes CLIP_BOUNDARY_FRAME D1/D3)
The Sept 3 boundary fix (half-frame epsilon around clip edges) made the
compositor disagree with the transport: an incoming clip could render up to
half a frame before its start and an outgoing clip linger half a frame past
its end, during playback and export alike. Peter saw "both clips or the
wrong clip edge showing."

New rule (Premiere/Resolve): a clip is active where start <= beat < end,
exactly — stopped, scrubbing, playing, exporting. The start edge belongs to
the clip; the end edge belongs to whatever follows (next clip, lower
layers, or black).

- Remove the epsilon query variants, visual_boundary_epsilon, and
  is_boundary_owned; restore the pre-epsilon exact collect (non-overlap
  makes tie-breaking unneeded).
- Min-remaining 20ms warm-up guard now applies to live playback only;
  stopped/scrub/export always start a clip the timeline says is active, so
  a final frame can be inspected and encoded.
- Export samples frame k at export_start + k*dt: the engine gets an
  absolute export clock (origin + frame_count * fixed_dt) instead of
  accumulating dt — accumulation drifted cuts off exact boundaries (a cut
  due at t=2.0s landed one frame late in the journey proof). Frame 0 does
  not integrate dt-driven effect state.
- ExportConfig beat fields are f64 end-to-end; the f32 round-trip could
  land the export start a hair outside a clip edge.

Verification: new journey proof export_cut_lands_exactly_on_clip_boundary
exports a two-layer cut fixture and asserts the visible cut lands on
exactly the boundary frame (caught the accumulation off-by-one; green
after the absolute clock). Boundary unit tests rewritten to the new rule.
1174 tests pass across core/playback/media/app; clippy -D warnings clean.

k3 (lead)

+++ crates/manifold-media/src/export_config.rs
@@ -14,13 +14,16 @@ pub struct ExportConfig {
     /// HDR export (HEVC 10-bit) vs SDR (H.264).
     pub hdr: bool,
     /// Export range start beat. 0.0 = use content range.
-    pub start_beat: f32,
+    /// f64: the project's beat domain is f64, and an f32 round-trip here can
+    /// land the export start a hair outside a clip edge, shifting which clip
+    /// owns frame 0 under exact boundary sampling.
+    pub start_beat: f64,
     /// Export range end beat. 0.0 = use content range.
-    pub end_beat: f32,
+    pub end_beat: f64,
     /// Optional audio file path for post-mux.
     pub audio_path: Option<String>,
     /// Beat position where the audio starts on the timeline.
-    pub audio_start_beat: f32,
+    pub audio_start_beat: f64,
     /// Encoder delay compensation in seconds (e.g. 0.05).
     pub audio_encoder_delay: f32,
     /// Split the export into one file per marker inside the export range.
